Fix spring-boot parent reference. Rename all tests that need to run when "mvn test" is executed.
