+++ pom.xml
@@ -3,8 +3,8 @@
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.springframework.boot</groupId>
-		<artifactId>spring-boot-starters</artifactId>
-		<version>1.2.0.RELEASE</version>
+		<artifactId>spring-boot-starter-parent</artifactId>
+		<version>1.2.3.RELEASE</version>
 	</parent>
 	<artifactId>mlpnn-experiment</artifactId>
 	<name>mlpnn</name>
@@ -40,7 +40,7 @@
 					<includes>
 						<include>**/*Test*.java</include>
 					</includes>
-					<argLine>-Xms512m -Xmx1024m -XX:MaxPermSize=512m</argLine>
+					<!--					<argLine>-Xms512m -Xmx1024m -XX:MaxPermSize=512m</argLine>-->
 				</configuration>
 			</plugin>
 			<plugin>
@@ -49,8 +49,8 @@
                 
 				<configuration>
 					<jvmArguments>
-						-Xms2048m -Xmx4096m -XX:MaxPermSize=2048m
-<!--						-Xdebug -Xrunjdwp:transport=dt_socket,server=y,suspend=y,address=5005-->
+						<!---Xms2048m -Xmx2048m -XX:MaxPermSize=2048m-->
+						<!--						-Xdebug -Xrunjdwp:transport=dt_socket,server=y,suspend=y,address=5005-->
 					</jvmArguments>
 				</configuration>
                 
@@ -74,7 +74,11 @@
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-thymeleaf</artifactId>
-		</dependency>        
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter-test</artifactId>
+		</dependency>		       
 		<dependency>
 			<groupId>org.springframework</groupId>
 			<artifactId>spring-core</artifactId>
